Overhaul the test runner
This allows us to catch up to recent changes in https://github.com/web-platform-tests/wpt/pull/39203. To reduce the maintenance burden, this rewrites the test runner to be focused on, as much as possible, reusing upstream runner code. For three files this is not possible, and those are explicitly called out.